pyproject.toml
[project]
name = "forge"
version = "0.1.0"
description = "AI-powered terminal for coding — local (Ollama) and cloud (Gemini) models"
requires-python = ">=3.11"
dependencies = [
    "textual>=3.0",
    "ollama>=0.4",
    "google-generativeai>=0.8",
]

[project.scripts]
forge = "forge.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
